pyproject.toml
[project]
name = "search-v2"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "annotated-types==0.7.0",
    "anyio==4.13.0",
    "certifi==2026.2.25",
    "cffi==2.0.0",
    "charset-normalizer==3.4.7",
    "cryptography==46.0.7",
    "distro==1.9.0",
    "doclayout-yolo>=0.0.4",
    "events==0.5",
    "exceptiongroup==1.3.1",
    "google-auth==2.49.2",
    "google-genai==1.73.1",
    "grpcio==1.80.0",
    "h11==0.16.0",
    "httpcore==1.0.9",
    "httpx==0.28.1",
    "huggingface-hub>=1.11.0",
    "idna==3.11",
    "numpy==2.2.6",
    "opensearch-protobufs==0.19.0",
    "opensearch-py==3.1.0",
    "pandas==2.3.3",
    "protobuf==7.34.1",
    "pyasn1==0.6.3",
    "pyasn1-modules==0.4.2",
    "pycparser==3.0",
    "pydantic==2.13.2",
    "pydantic-core==2.46.2",
    "pymupdf==1.27.2.2",
    "python-dateutil==2.9.0.post0",
    "python-dotenv==1.2.2",
    "pytz==2026.1.post1",
    "requests==2.33.1",
    "sentence-transformers>=5.4.1",
    "six==1.17.0",
    "sniffio==1.3.1",
    "tenacity==9.1.4",
    "tqdm==4.67.3",
    "typing-extensions==4.15.0",
    "typing-inspection==0.4.2",
    "tzdata==2026.1",
    "ultralytics>=8.4.41",
    "urllib3==2.6.3",
    "websockets==16.0",
]
